Add DownloadButton component and enhance error handling in user and stock actions

+++ actions/user/destroyUser.ts
@@ -4,6 +4,16 @@ import prisma from "@/lib/prisma"
 
 export const destroyUser = async (id: number) => {
   try {
+    const existUser = await prisma.users.findFirst({
+      where: {
+        id: id
+      }
+    })
+
+    if (!existUser) {
+      throw new Error('El usuario no existe')
+    }
+
     const user = await prisma.users.delete({
       where: {
         id: id
@@ -12,6 +22,10 @@ export const destroyUser = async (id: number) => {
 
     return true
   } catch (error: any) {
-    throw new Error(error)
+    if (error.message.includes('El usuario no existe')) {
+      throw new Error(error.message)
+    } else {
+      throw new Error('Error al eliminar el usuario')
+    }
   }
 }
